# public/docs/openapi.yaml
openapi: 3.0.3
info:
  title: Laravel
  description: ''
  version: 1.0.0
servers:
  -
    url: 'http://localhost'
paths:
...
  /api/cart/add-product:
    post:
      summary: 'Add Product'
      operationId: addProduct
      description: 'Add a product to cart.'
      parameters: []
      responses: {  }
      tags:
        - 'Cart management'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                product_id:
                  type: integer
                  description: ''
                  example: 14
                quantity:
                  type: number
                  description: ''
                  example: 1.53273
                is_unit:
                  type: boolean
                  description: ''
                  example: true
              required:
                - product_id
                - quantity
                - is_unit
      security: []
...